// frontend/src/app/models/korisnik.ts
export class Korisnik{
    id: number;
    ime: string;
    prezime: string;
    kor_ime: string;
    lozinka: string;
    tip: string;
    mejl: string;
    telefon: string;
    naziv: string;
    adresa: string;
    maticni_broj: string;
    opis: string;
    slobodno: number;
    zahtev: boolean;
    slika: string;
    dodatno: number;
}